move BgpConfig from common to nexus-types
